III-3128 Add separate culturefeed service
